+++ cmd/birdy/adopt_other.go
@@ -1,0 +1,7 @@
+//go:build !unix
+
+package main
+
+// adoptDBOwnership is a Unix concern: there is no service user to hand the
+// database to on other platforms.
+func adoptDBOwnership(string) {}